+++ bocadillo/applications.py
@@ -44,7 +44,7 @@ from .middleware import ASGIMiddleware
 from .request import Request
 from .response import Response
 from .routing import RoutingMixin
-from .staticfiles import static
+from .staticfiles import static, WhiteNoise
 from .templates import TemplatesMixin
 
 if TYPE_CHECKING:  # pragma: no cover
@@ -75,6 +75,9 @@ class App(TemplatesMixin, RoutingMixin, metaclass=DocsMeta):
     static_root (str):
         The path prefix for static assets.
         Defaults to `"static"`.
+    static_config (dict):
+        Extra static files configuration attributes.
+        See also [static](./staticfiles.md#static).
     allowed_hosts (list of str, optional):
         A list of hosts which the server is allowed to run at.
         If the list contains `"*"`, any host is allowed.
@@ -118,6 +121,7 @@ class App(TemplatesMixin, RoutingMixin, metaclass=DocsMeta):
         *,
         static_dir: Optional[str] = "static",
         static_root: Optional[str] = "static",
+        static_config: dict = None,
         allowed_hosts: List[str] = None,
         enable_cors: bool = False,
         cors_config: dict = None,
@@ -140,6 +144,7 @@ class App(TemplatesMixin, RoutingMixin, metaclass=DocsMeta):
         # Mounted (children) apps
         self._prefix_to_app: Dict[str, Any] = {}
         self._name_to_prefix_and_app: Dict[str, Tuple[str, App]] = {}
+        self._static_apps: Dict[str, WhiteNoise] = {}
 
         # Test client
         self.client = self.build_client()
@@ -148,7 +153,7 @@ class App(TemplatesMixin, RoutingMixin, metaclass=DocsMeta):
         if static_dir is not None:
             if static_root is None:
                 static_root = static_dir
-            self.mount(static_root, static(static_dir))
+            self.mount(static_root, static(static_dir, **(static_config or {})))
 
         # Media
         self.media_handlers = get_default_handlers()
@@ -254,6 +259,9 @@ class App(TemplatesMixin, RoutingMixin, metaclass=DocsMeta):
 
         if isinstance(app, App) and app.name is not None:
             self._name_to_prefix_and_app[app.name] = (prefix, app)
+
+        if isinstance(app, WhiteNoise):
+            self._static_apps[prefix] = app
 
     def recipe(self, recipe: "Recipe"):
         """Apply a recipe.
@@ -439,7 +447,10 @@ class App(TemplatesMixin, RoutingMixin, metaclass=DocsMeta):
         - [Uvicorn settings](https://www.uvicorn.org/settings/) for all
         available keyword arguments.
         """
+        live_server = False
+
         if _run is None:  # pragma: no cover
+            live_server = True
             _run = run
 
         if "PORT" in os.environ:
@@ -453,9 +464,18 @@ class App(TemplatesMixin, RoutingMixin, metaclass=DocsMeta):
         if port is None:
             port = 8000
 
-        if debug:
-            self.debug = True
-            reloader = StatReload(get_logger(log_level))
+        if not debug:
+            _run(self, host=host, port=port, **kwargs)
+            return
+
+        self.debug = True
+
+        # Reload static files in development.
+        # See: http://whitenoise.evans.io/en/stable/base.html#autorefresh
+        for whitenoise in self._static_apps.values():
+            whitenoise.autorefresh = True
+
+        if live_server:
             kwargs = {
                 "app": self,
                 "host": host,
@@ -464,7 +484,8 @@ class App(TemplatesMixin, RoutingMixin, metaclass=DocsMeta):
                 "debug": self.debug,
                 **kwargs,
             }
-            reloader.run(run, kwargs)
+            reloader = StatReload(get_logger(log_level))
+            reloader.run(_run, kwargs)
         else:
             _run(self, host=host, port=port, **kwargs)